+++ src/org/ojalgo/matrix/geometry/Primitive32Vector2.java
@@ -24,10 +24,10 @@ public class Primitive32Vector2 extends GeometryVector {
     public final void add(final int row, final double addend) {
         switch (row) {
         case 0:
-            v0 += addend;
+            v0 += (float) addend;
             break;
         case 1:
-            v1 += addend;
+            v1 += (float) addend;
             break;
         default:
             throw new IllegalArgumentException();
